fix(crypto): checked usize conversion for PBKDF2 length (review, PR #607)
`length` was cast `u64 as usize` at the call sites, which truncates on 32-bit
targets (e.g. wasm32): a value above usize::MAX could wrap to a small number and
slip past the MAX_PBKDF2_KEY_LENGTH check, deriving a wrong-length key instead of
erroring. Thread the raw u64 into pbkdf2_hmac_sha256_str and convert once with a
checked usize::try_from that returns a clear RuntimeError on overflow. No change
on 64-bit; existing KDF vector tests pass.

+++ src/stdlib/crypto_async.rs
@@ -113,7 +113,9 @@ fn pbkdf2_hmac_route(args: &[Value]) -> RoutedFuture {
         let password = Zeroizing::new(expect_text(&args[0])?.to_string());
         let salt = expect_text(&args[1])?.to_string();
         let iterations = crypto::expect_count(FUNC, "iterations", &args[2])?;
-        let length = crypto::expect_count(FUNC, "length", &args[3])? as usize;
+        // Pass the raw u64 through; `pbkdf2_hmac_sha256_str` does a checked
+        // usize conversion so a value that doesn't fit can't truncate.
+        let length = crypto::expect_count(FUNC, "length", &args[3])?;
         Ok::<_, RuntimeError>((password, salt, iterations, length))
     })();
     Box::pin(async move {
